tests/wpt/tests/css/css-sizing/contain-intrinsic-size/parsing/from-element-valid.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Intrinsic Size Test: intrinsic-size with from-element</title>
<link rel="author" title="Chris Harrelson" href="mailto:chrishtr@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#intrinsic-size-override">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("contain-intrinsic-width", "from-element none");
test_valid_value("contain-intrinsic-height", "from-element none");

test_valid_value("contain-intrinsic-inline-size", "from-element none");
test_valid_value("contain-intrinsic-block-size", "from-element none");

test_valid_value("contain-intrinsic-width", "from-element 200px");
test_valid_value("contain-intrinsic-height", "from-element 200px");

test_valid_value("contain-intrinsic-size", "from-element none");
test_valid_value("contain-intrinsic-size", "from-element 200px");
test_valid_value("contain-intrinsic-size", "from-element 300px 200px");
test_valid_value("contain-intrinsic-size", "300px from-element 200px");
test_shorthand_value("contain-intrinsic-size", "from-element none", {
  "contain-intrinsic-width": "from-element none",
  "contain-intrinsic-height": "from-element none",
});
test_shorthand_value("contain-intrinsic-size", "from-element 200px", {
  "contain-intrinsic-width": "from-element 200px",
  "contain-intrinsic-height": "from-element 200px",
});
test_shorthand_value("contain-intrinsic-size", "from-element 300px 200px", {
  "contain-intrinsic-width": "from-element 300px",
  "contain-intrinsic-height": "200px",
});
test_shorthand_value("contain-intrinsic-size", "300px from-element 200px", {
  "contain-intrinsic-width": "300px",
  "contain-intrinsic-height": "from-element 200px",
});
</script>
</body>
</html>
